refactor: decrease code duplication
