+++ Expressif.Cli/Commands/EvaluateCommand.cs
@@ -1,5 +1,4 @@
 using System.CommandLine;
-using System.Text;
 
 namespace Expressif.Cli.Commands;
 
@@ -42,35 +41,17 @@ internal static class EvaluateCommand
 
         command.SetAction(parseResult =>
         {
-            var expressionCode = parseResult.GetValue(expressionArgument);
+            var inlineExpression = parseResult.GetValue(expressionArgument);
             var expressionFilePath = parseResult.GetValue(expressionFileOption);
             var input = parseResult.GetValue(inputOption);
             var hasInputOption = parseResult.GetResult(inputOption) is not null;
 
-            var hasInlineExpression = !string.IsNullOrWhiteSpace(expressionCode);
-            var hasExpressionFile = !string.IsNullOrWhiteSpace(expressionFilePath);
-
-            if (hasInlineExpression && hasExpressionFile)
+            if (!ExpressionCommandCommon.TryResolveExpressionCode(
+                    inlineExpression,
+                    expressionFilePath,
+                    out var expressionCode,
+                    out var hasExpressionFile))
             {
-                Console.Error.WriteLine("The expression cannot be provided both inline and through --file.");
-                return ExitCodes.InvalidExpressionOrInput;
-            }
-
-            if (!hasInlineExpression && !hasExpressionFile)
-            {
-                Console.Error.WriteLine("The expression must be supplied through exactly one source: inline or --file.");
-                return ExitCodes.InvalidExpressionOrInput;
-            }
-
-            if (hasExpressionFile
-                && !TryReadExpressionFile(expressionFilePath!, out expressionCode))
-            {
-                return ExitCodes.InvalidExpressionOrInput;
-            }
-
-            if (string.IsNullOrWhiteSpace(expressionCode))
-            {
-                Console.Error.WriteLine("Expression is required.");
                 return ExitCodes.InvalidExpressionOrInput;
             }
 
@@ -92,14 +73,7 @@ internal static class EvaluateCommand
                                                   or NotImplementedFunctionException
                                                   or MissingOrUnexpectedParametersFunctionException)
                 {
-                    if (hasExpressionFile)
-                    {
-                        Console.Error.WriteLine($"The expression loaded from '{expressionFilePath}' is invalid:");
-                        Console.Error.WriteLine(CommandErrorFormatter.FormatValidationError(exception));
-                        return ExitCodes.InvalidExpressionOrInput;
-                    }
-
-                    return CommandErrorFormatter.WriteValidationError(exception);
+                    return ExpressionCommandCommon.WriteValidationError(exception, hasExpressionFile, expressionFilePath);
                 }
                 catch (Exception exception)
                 {
@@ -134,14 +108,7 @@ internal static class EvaluateCommand
                                               or NotImplementedFunctionException
                                               or MissingOrUnexpectedParametersFunctionException)
             {
-                if (hasExpressionFile)
-                {
-                    Console.Error.WriteLine($"The expression loaded from '{expressionFilePath}' is invalid:");
-                    Console.Error.WriteLine(CommandErrorFormatter.FormatValidationError(exception));
-                    return ExitCodes.InvalidExpressionOrInput;
-                }
-
-                return CommandErrorFormatter.WriteValidationError(exception);
+                return ExpressionCommandCommon.WriteValidationError(exception, hasExpressionFile, expressionFilePath);
             }
             catch (Exception exception)
             {
@@ -163,47 +130,5 @@ internal static class EvaluateCommand
         });
 
         return command;
-    }
-
-    private static bool TryReadExpressionFile(string path, out string expressionCode)
-    {
-        expressionCode = string.Empty;
-
-        if (Directory.Exists(path))
-        {
-            Console.Error.WriteLine($"Expression file '{path}' is a directory.");
-            return false;
-        }
-
-        if (!File.Exists(path))
-        {
-            Console.Error.WriteLine($"Expression file '{path}' was not found.");
-            return false;
-        }
-
-        try
-        {
-            using var stream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.Read);
-            using var reader = new StreamReader(stream, new UTF8Encoding(encoderShouldEmitUTF8Identifier: false, throwOnInvalidBytes: true), detectEncodingFromByteOrderMarks: true);
-            expressionCode = reader.ReadToEnd();
-        }
-        catch (DecoderFallbackException)
-        {
-            Console.Error.WriteLine($"Expression file '{path}' could not be decoded as UTF-8.");
-            return false;
-        }
-        catch (Exception exception) when (exception is IOException or UnauthorizedAccessException)
-        {
-            Console.Error.WriteLine($"Expression file '{path}' could not be accessed: {exception.Message}");
-            return false;
-        }
-
-        if (string.IsNullOrWhiteSpace(expressionCode))
-        {
-            Console.Error.WriteLine($"Expression file '{path}' is empty.");
-            return false;
-        }
-
-        return true;
     }
 }
